initial commit
"matrix.py" is an easy to use python3.5 program for plotting and working with matricies and vectors. It uses the matplotlib library for plotting, and overloads many operations for ease. Please note: the code for finding matrix identities (lines 203 through 241) were 
adapted from another open-source site. A complete documentation will be posted soon (still working on it).